implemented feedback
