+++ src/clients/sqlite-client.ts
@@ -1,0 +1,6 @@
+import { SqliteClientProxyOpfs } from "./sqlite-client-proxy-opfs";
+
+// backwards compatibility
+export class SqliteClient extends SqliteClientProxyOpfs {
+
+};
