+++ src/enums/sqlite-client-type.enum.ts
@@ -1,0 +1,6 @@
+export enum SqliteClientType {
+    Opfs = "OPFS",
+    OpfsSah = "OPFS_SAH",
+    MemoryMain = "MEMORY_MAIN",
+    MemoryWorker = "MEMORY_WORKER"
+}
