+++ src/enums/sqlite-message-type.enum.ts
@@ -1,9 +1,3 @@
-export enum SqliteClientType {
-  Opfs = "OPFS",
-  OpfsSah = "OPFS_SAH",
-  Memory = "MEMORY"
-}
-
 export enum SqliteMessageTypeEnum {
   CreateDatabase = "CREATE_DATABASE",
   CreateDatabaseResult = "CREATE_DATABASE_RESULT",
