+++ src/interfaces/sqlite-client-common-options.interface.ts
@@ -1,0 +1,4 @@
+export interface SqliteClientCommonOptions {
+    filename: string,
+  }
+
