+++ src/interfaces/sqlite-client-memory-options.interface.ts
@@ -1,0 +1,5 @@
+import { SqliteClientCommonOptions } from "./sqlite-client-common-options.interface";
+
+export interface SqliteClientMemoryOptions extends SqliteClientCommonOptions {
+  flags: string,
+}
