+++ src/interfaces/sqlite-client-proxy-memory-options.interface.ts
@@ -1,0 +1,5 @@
+import { SqliteClientProxyOptions } from "./sqlite-client-proxy-options.interface";
+
+export interface SqliteClientProxyMemoryOptions extends SqliteClientProxyOptions {
+  flags: string,
+}
