+++ src/interfaces/sqlite-client-proxy-opfs-options.interface.ts
@@ -1,0 +1,5 @@
+import { SqliteClientProxyOptions } from "./sqlite-client-proxy-options.interface";
+
+export interface SqliteClientProxyOpfsOptions extends SqliteClientProxyOptions {
+  flags: string,
+}
