+++ src/interfaces/sqlite-client-proxy-options.interface.ts
@@ -1,0 +1,6 @@
+import { SqliteClientCommonOptions } from "./sqlite-client-common-options.interface";
+  
+  export interface SqliteClientProxyOptions extends SqliteClientCommonOptions {
+    sqliteWorkerPath: string,
+  }
+
