+++ src/messages/create-database-result.message.ts
@@ -1,5 +1,5 @@
-import {SqliteMessageInterface} from "../interfaces/sqlite-message.interface";
-import {SqliteMessageTypeEnum} from "../enums/sqlite-message-type.enum";
+import { SqliteMessageInterface } from "../interfaces/sqlite-message.interface";
+import { SqliteMessageTypeEnum } from "../enums/sqlite-message-type.enum";
 
 export class CreateDatabaseResultMessage implements SqliteMessageInterface {
   type: SqliteMessageTypeEnum = SqliteMessageTypeEnum.CreateDatabaseResult;
