+++ src/messages/execute-sql-result.message.ts
@@ -1,12 +1,11 @@
-import {SqliteMessageInterface} from "../interfaces/sqlite-message.interface";
-import {SqliteMessageTypeEnum} from "../enums/sqlite-message-type.enum";
-
+import { SqliteMessageInterface } from "../interfaces/sqlite-message.interface";
+import { SqliteMessageTypeEnum } from "../enums/sqlite-message-type.enum";
 export class ExecuteSqlResultMessage implements SqliteMessageInterface {
   type: SqliteMessageTypeEnum = SqliteMessageTypeEnum.ExecuteSqlResult;
 
   constructor(public readonly uniqueId: string,
-              public readonly result: any,
-              public readonly error?: any) {
+    public readonly result: any,
+    public readonly error?: any) {
   }
 
 }
